introduced parent class PBEIOBase to unify PBSIO and SGEIO functionalities and similarities
