+++ src/qtoolkit/io/sge.py
@@ -3,26 +3,10 @@ from __future__ import annotations
 import re
 import xml.dom.minidom
 import xml.parsers.expat
-from datetime import timedelta
 
-from qtoolkit.core.data_objects import (
-    CancelResult,
-    CancelStatus,
-    ProcessPlacement,
-    QJob,
-    QJobInfo,
-    QResources,
-    QState,
-    QSubState,
-    SubmissionResult,
-    SubmissionStatus,
-)
-from qtoolkit.core.exceptions import (
-    CommandFailedError,
-    OutputParsingError,
-    UnsupportedResourcesError,
-)
-from qtoolkit.io.base import BaseSchedulerIO
+from qtoolkit.core.data_objects import QJob, QJobInfo, QResources, QState, QSubState
+from qtoolkit.core.exceptions import CommandFailedError, OutputParsingError
+from qtoolkit.io.pbs_base import PBSIOBase
 
 # https://wiki.nikhil.io/Ancient_Sysadmin_Stuff/Sun_Grid_Engine_States/
 # https://manpages.ubuntu.com/manpages/jammy/en/man5/sge_status.5.html
@@ -118,7 +102,7 @@ _STATUS_MAPPING = {
 }
 
 
-class SGEIO(BaseSchedulerIO):
+class SGEIO(PBSIOBase):
     header_template: str = """
 #$ -cwd $${cwd}
 #$ -q $${queue}
@@ -145,64 +129,17 @@ $${qverbatim}"""
     def __init__(self, get_job_executable: str = "qstat"):
         self.get_job_executable = get_job_executable
 
-    def parse_submit_output(self, exit_code, stdout, stderr) -> SubmissionResult:
-        if isinstance(stdout, bytes):
-            stdout = stdout.decode()
-        if isinstance(stderr, bytes):
-            stderr = stderr.decode()
-        if exit_code != 0:
-            return SubmissionResult(
-                exit_code=exit_code,
-                stdout=stdout,
-                stderr=stderr,
-                status=SubmissionStatus("FAILED"),
-            )
+    def extract_job_id(self, stdout):
         match = re.search(r'Your job (\d+) \(".*?"\) has been submitted', stdout)
         if not match:
             raise OutputParsingError("Failed to parse job ID from stdout")
-        job_id = match.group(1)
-        status = (
-            SubmissionStatus("SUCCESSFUL")
-            if job_id
-            else SubmissionStatus("JOB_ID_UNKNOWN")
-        )
-        return SubmissionResult(
-            job_id=job_id,
-            exit_code=exit_code,
-            stdout=stdout,
-            stderr=stderr,
-            status=status,
-        )
+        return match.group(1)
 
-    def parse_cancel_output(self, exit_code, stdout, stderr) -> CancelResult:
-        """Parse the output of the qdel command."""
-        if isinstance(stdout, bytes):
-            stdout = stdout.decode()
-        if isinstance(stderr, bytes):
-            stderr = stderr.decode()
-        if exit_code != 0:
-            return CancelResult(
-                exit_code=exit_code,
-                stdout=stdout,
-                stderr=stderr,
-                status=CancelStatus("FAILED"),
-            )
+    def extract_job_id_from_cancel(self, stderr):
         match = re.search(r"qdel: job (\d+) deleted", stderr)
         if not match:
             raise OutputParsingError("Failed to parse job ID from stdout")
-        job_id = match.group(1)
-        status = CancelStatus("SUCCESSFUL")
-        return CancelResult(
-            job_id=job_id,
-            exit_code=exit_code,
-            stdout=stdout,
-            stderr=stderr,
-            status=status,
-        )
-
-    def _get_job_cmd(self, job_id: str):
-        cmd = f"qstat -j {job_id}"
-        return cmd
+        return match.group(1)
 
     def parse_job_output(self, exit_code, stdout, stderr) -> QJob | None:  # aiida style
         if exit_code != 0:
@@ -408,133 +345,23 @@ $${qverbatim}"""
             raise OutputParsingError(f"Invalid time format: {time_str}")
 
     @staticmethod
-    def _convert_memory_str(memory: str | None) -> int | None:
-        if not memory:
-            return None
-
-        match = re.match(r"([0-9]+)([a-zA-Z]*)", memory)
-        if not match:
-            raise OutputParsingError("No numbers and units parsed")
-        memory, units = match.groups()
-
-        power_labels = {"K": 0, "M": 1, "G": 2, "T": 3}
-
-        if not units:
-            units = "M"
-        elif units not in power_labels:
-            raise OutputParsingError(f"Unknown units {units}")
-        try:
-            v = int(memory)
-        except ValueError:
-            raise OutputParsingError
-
-        return v * (1024 ** power_labels[units.upper()])
-
-    _qresources_mapping = {
-        "queue_name": "queue",
-        "job_name": "job_name",
-        "priority": "priority",
-        "output_filepath": "qout_path",
-        "error_filepath": "qerr_path",
-        "project": "group_list",
-    }
+    def get_power_labels() -> dict:
+        return {"k": 0, "m": 1, "g": 2, "t": 3}
 
     @staticmethod
-    def _convert_time_to_str(time: int | float | timedelta) -> str:
-        if not isinstance(time, timedelta):
-            time = timedelta(seconds=time)
+    def get_default_unit() -> str:
+        return "M"
 
-        hours, remainder = divmod(int(time.total_seconds()), 3600)
-        minutes, seconds = divmod(remainder, 60)
+    def get_system_name(self) -> str:
+        return "SGE"
 
-        time_str = f"{hours}:{minutes}:{seconds}"
-        return time_str
-
-    def _convert_qresources(self, resources: QResources) -> dict:
-        header_dict = {}
-        for qr_field, sge_field in self._qresources_mapping.items():
-            val = getattr(resources, qr_field)
-            if val is not None:
-                header_dict[sge_field] = val
-
-        if resources.njobs and resources.njobs > 1:
-            header_dict["array"] = f"1-{resources.njobs}"
-
-        if resources.time_limit:
-            header_dict["walltime"] = self._convert_time_to_str(resources.time_limit)
-            header_dict["soft_walltime"] = self._convert_time_to_str(
-                resources.time_limit * 0.99
-            )
-
-        if resources.rerunnable is not None:
-            header_dict["rerunnable"] = "y" if resources.rerunnable else "n"
-
-        nodes, processes, processes_per_node = resources.get_processes_distribution()
-        select = None
-        if resources.process_placement == ProcessPlacement.NO_CONSTRAINTS:
-            select = f"select={processes}"
-            if resources.threads_per_process:
-                select += f":ncpus={resources.threads_per_process}"
-                select += f":ompthreads={resources.threads_per_process}"
-            if resources.memory_per_thread:
-                threads_per_process = resources.threads_per_process or 1
-                select += f":mem={threads_per_process * resources.memory_per_thread}mb"
-        elif resources.process_placement in (
-            ProcessPlacement.EVENLY_DISTRIBUTED,
-            ProcessPlacement.SAME_NODE,
-            ProcessPlacement.SCATTERED,
-        ):
-            select = f"select={nodes}"
-            if resources.threads_per_process and resources.threads_per_process > 1:
-                cpus = resources.threads_per_process * processes_per_node
-                ompthreads = resources.threads_per_process
-            else:
-                cpus = processes_per_node
-                ompthreads = None
-            select += f":ncpus={cpus}"
-            select += f":mpiprocs={processes_per_node}"
-            if ompthreads:
-                select += f":ompthreads={ompthreads}"
-            if resources.memory_per_thread:
-                mem = cpus * resources.memory_per_thread
-                select += f":mem={mem}mb"
-
-            if resources.process_placement in (
-                ProcessPlacement.EVENLY_DISTRIBUTED,
-                ProcessPlacement.SCATTERED,
-            ):
-                header_dict["place"] = "scatter"
-            elif resources.process_placement == ProcessPlacement.SAME_NODE:
-                header_dict["place"] = "pack"
-        else:
-            msg = f"process placement {resources.process_placement} is not supported for SGE"
-            raise UnsupportedResourcesError(msg)
-
-        header_dict["select"] = select
-
-        if resources.email_address:
-            header_dict["mail_user"] = resources.email_address
-            header_dict["mail_type"] = "abe"
-
-        if resources.scheduler_kwargs:
-            header_dict.update(resources.scheduler_kwargs)
-
-        return header_dict
+    def _add_soft_walltime(self, header_dict: dict, resources: QResources):
+        header_dict["soft_walltime"] = self._convert_time_to_str(
+            resources.time_limit * 0.99
+        )
 
     @property
     def supported_qresources_keys(self) -> list:
-        supported = list(self._qresources_mapping.keys())
-        supported += [
-            "njobs",
-            "memory_per_thread",
-            "time_limit",
-            "processes",
-            "processes_per_node",
-            "process_placement",
-            "nodes",
-            "threads_per_process",
-            "email_address",
-            "scheduler_kwargs",
-            "gpus_per_job",
-        ]
+        supported = super().supported_qresources_keys
+        supported += ["memory_per_thread", "gpus_per_job"]
         return supported
